// types/platform.types.ts
export interface SpotifyStatus {
  connected: boolean;
  displayName: string | null;
}

export interface YoutubeStatus {
  connected: boolean;
  channelTitle: string | null;
}

export interface DeezerStatus {
  connected: boolean;
  displayName: string | null;
}

export interface AppleStatus {
  connected: boolean;
  displayName: string | null;
}